Refactor: move GeohashUtils to 'io' package since it is after all an encoding format.  No need for 'util' packabge anymore.
